Increased water consumption for sea dragon in desert
Sea dragon now consumes more water in biomes like the desert, nether etc (Any biomes that has no rain and BaseTemperature above 1.0). Water consumption is also halved during night time.

+++ src/main/java/by/jackraidenph/dragonsurvival/handlers/SpecificsHandler.java
@@ -193,7 +193,7 @@ public class SpecificsHandler {
     				ForgeIngameGui.right_height += 10;
 
     				int maxTimeWithoutWater = ConfigHandler.SERVER.seaTicksWithoutWater.get();
-    				int timeWithoutWater = maxTimeWithoutWater - playerStateHandler.getDebuffData().timeWithoutWater;
+				    double timeWithoutWater = maxTimeWithoutWater - playerStateHandler.getDebuffData().timeWithoutWater;
     				boolean flag = false;
     				if (timeWithoutWater < 0) {
     					flag = true;
